Add support for TCP config phase, double precision euler
Up-to-date with https://github.com/michprev/flyhero-esp32/commit/2a8ede6a871343dc19e5fa5ecea844f0b49f5868

+++ src/shared_data.h
@@ -4,16 +4,19 @@
 
 #pragma once
 
-struct Euler_Data {
-    float roll, pitch, yaw;
+struct Euler_Data
+{
+    double roll, pitch, yaw;
 
-    Euler_Data() {
+    Euler_Data()
+    {
         this->roll = 0;
         this->pitch = 0;
         this->yaw = 0;
     }
 
-    Euler_Data(float roll, float pitch, float yaw) {
+    Euler_Data(double roll, double pitch, double yaw)
+    {
         this->roll = roll;
         this->pitch = pitch;
         this->yaw = yaw;
